python-webgame/requirements.txt
flask
wtforms
flask_wtf
mysql-connector
